統一密碼驗證由一檔案控制
