+++ src/components/registerDialog.vue
@@ -69,6 +69,7 @@ import { ref, reactive, inject } from 'vue'
 import { useI18n } from 'vue-i18n'
 import { useUserStore } from 'src/stores/user'
 import notify from '../utils/notify'
+import valList from '../utils/data/valList'
 const isPwd = ref(true)
 const { t } = useI18n()
 const users = useUserStore()
@@ -99,26 +100,7 @@ const stepper = ref(null)
 const mailVerifying = ref(false)
 const mailCodeValid = ref(null)
 //
-const emailVal = (isSchool) => {
-  const rule = [
-    val => (val !== null && val !== '') || 'Please type your email',
-    val => val.length <= 60 || '必須 60 個字以下'
-  ]
-  if (isSchool) {
-    rule.push(
-      // @後方必須含 .edu.
-      // val => val.match(/^[a-z0-9]+@[a-z0-9\.]+\.edu\.[a-z0-9\.]+$/) || '格式錯誤/必須為學校信箱')
-      // 初期限制要有".edu.tw"結尾才行
-      // eslint-disable-next-line no-useless-escape
-      val => (/^[a-z0-9]+@[a-z0-9\.]+\.edu\.tw$/).test(val) || '格式錯誤，僅可含英小寫、數、@、./必須為學校信箱')
-  } else {
-    rule.push(
-      // eslint-disable-next-line no-useless-escape
-      val => (/^[a-z0-9]+@[a-z0-9]+\.[a-z0-9\.]+$/).test(val) || '格式錯誤，僅可含英小寫、數、@、.'
-    )
-  }
-  return rule
-}
+const emailVal = valList.mail
 const nextPage = async () => {
   try {
     if (step.value === 2) {
@@ -140,23 +122,11 @@ const nextPage = async () => {
 const registerForm = reactive({ schoolEmail: '', schoolEmailCode: '', account: '', password: '', nickName: '', gender: '0' })
 
 // ***********rule val區******************************
-const accountVal = [
-  val => (val && val.length >= 6 && val.length <= 30) || '長度需介於6~30字之間',
-  val => (/^[a-zA-Z0-9]+$/).test(val) || '只能輸入英文與數字'
-]
-const passwordVal = [
-  val => (val && val.length >= 8 && val.length <= 40) || '長度需介於8~40字之間',
-  val => ((/[a-zA-Z]/).test(val) && (/[0-9]/).test(val)) || '必須含英文與數字',
-  val => true || '預留給有同名使用，名稱重複會把這行換成不能等於原本帳號'
-]
-const nickNameVal = [
-  val => (val && val.length >= 6 && val.length <= 30) || '長度需介於6~30字之間',
-  val => (!['owner', 'you', 'youHide', 'admin'].includes(val)) || '該暱稱不可使用',
-  val => true || '預留給有同名使用'
-]
-const mailCodeVal = [
-  val => (val.length === 10 && (/^[a-zA-Z0-9]+$/).test(val)) || '為十位英數'
-]
+const accountVal = valList.account
+const passwordVal = valList.password
+const nickNameVal = valList.nickName
+const mailCodeVal = valList.tenCode
+
 const register = async () => {
   try {
     const rep = await users.register(registerForm)
